[csi-driver,api] Rewrite CSI for RV and make RSC fields optional
CSI driver (rewrite for the new control plane):
- Rewrite the CSI driver against ReplicatedVolume and
  ReplicatedVolumeAttachment CRDs instead of ReplicatedStoragePool +
  LVM
- CreateVolume uses ReplicatedStorageClassName, supports
  WaitForFirstConsumer and Immediate binding (early RVA with rollback
  for WFFC), annotates the RV with PVC name/namespace, and rejects
  snapshot/clone requests
- DeleteVolume cleans up orphaned RVAs; ControllerExpandVolume uses a
  direct size comparison; drop the CLONE_VOLUME capability
- NodeGetInfo reports AccessibleTopology; NodeStageVolume resolves
  device symlinks and drops LVM params; set MaxAttachments to 2; add
  RVA RBAC and remove the unused StoragePool/LVM utilities

API status phase/message:
- Add phase and message status fields to RSC (7-value phase), RVR
  (12-value health model), and RVA (refined: drop the unreachable
  Detached phase, fix the message override, add an empty-message
  fallback)
- Improve DRBDResource printcolumns (observed state, Configured,
  Message)
- Rename Deleting to Terminating in phase values and condition reasons

API field defaulting (storage-optional):
- Make the RSC rollout strategy, conflict/eligible-node policies,
  systemNetworkNames, and storage optional pointers; fill defaults via
  a controller-side reconcileDefaults instead of CRD defaults
- Remove the CRD default from RSC spec.replication so FTT/GMDR are
  usable
- Make RV maxAttachments an optional pointer with a CRD default
- Fix CEL validation for the optional replicatedStorageClassName field

Controller:
- rv_controller: expand RVR event predicates to react to Quorum changes,
  Attachment changes (DevicePath, IOSuspended, InUse), the Ready
  condition, and peer BackingVolumeState and connection-establishment
  changes, so attachment and detach guards re-run promptly; add
  predicates_test.go

Webhooks (dual control plane):
- Detect new vs legacy control plane via the controller-config
  ConfigMap and validate RSC accordingly (new requires spec.storage;
  legacy requires spec.storagePool and rejects new-style fields)
- Add RBAC and a ValidatingWebhookConfiguration for RSC; return errors
  instead of klog.Fatal

Agent:
- Tear down orphaned on-node DRBD resources on force-delete; skip
  non-managed resources
- Respect maintenance mode in the rename and DRBDResourceOperation
  paths
- Parse rs-in-flight/ap-in-flight as uint64 to avoid an unmarshal
  overflow

Build and templates:
- Bump DRBD to 9.2.16-flant.2 and switch the DRBD source to
  deckhouse/3p-drbd; switch the agent image to a distroless base
- Remove the newControlPlane condition from the DRBD install
  NodeGroupConfiguration templates and comment out the agent PodMonitor

+++ images/agent/pkg/drbdutils/status.go
@@ -65,8 +65,8 @@ type Connection struct {
 	Congested       bool   `json:"congested"`
 	Peerrole        string `json:"peer-role"`
 	TLS             bool   `json:"tls"`
-	APInFlight      int    `json:"ap-in-flight"`
-	RSInFlight      int    `json:"rs-in-flight"`
+	APInFlight      uint64 `json:"ap-in-flight"`
+	RSInFlight      uint64 `json:"rs-in-flight"`
 
 	Paths       []Path       `json:"paths"`
 	PeerDevices []PeerDevice `json:"peer_devices"`
